Scan wren_lift's fiber stacks
wren_lift's seam tells the adapter when a krio stack is made, suspended
and freed; the adapter registers each with the heap. Stacks register
under krio's id, unique in the process, so the core's own tasks use it
too. The driver and the runner turn krio fibers on for the VMs they make.

git-bug: 172477ac3666987976e6dc0985796fca5fe2307c590300c616b5199919777d87

+++ crates/caribou-ash/src/heap.rs
@@ -85,15 +85,15 @@ pub unsafe extern "C" fn gc_unregister_current_os_thread() {
 }
 
 pub unsafe extern "C" fn gc_register_fiber_stack(id: u32, base: usize, size: usize) {
-    unsafe { heap::gc_register_fiber_stack(id, base, size) };
+    unsafe { heap::gc_register_fiber_stack(u64::from(id), base, size) };
 }
 
 pub unsafe extern "C" fn gc_update_fiber_sp(id: u32, sp: usize) {
-    unsafe { heap::gc_update_fiber_sp(id, sp) };
+    unsafe { heap::gc_update_fiber_sp(u64::from(id), sp) };
 }
 
 pub unsafe extern "C" fn gc_unregister_fiber_stack(id: u32) {
-    unsafe { heap::gc_unregister_fiber_stack(id) };
+    unsafe { heap::gc_unregister_fiber_stack(u64::from(id)) };
 }
 
 pub unsafe extern "C" fn gc_add_persistent(ptr: *mut vdynamic) {
